+++ packages/tts-editor/src/tts.d.ts
@@ -1,0 +1,2 @@
+/// <reference types="@typed-tabletop-simulator/declaration" />
+/// <reference types="@typescript-to-lua/language-extensions" />